lang-portal/opea-comps/mega-service/requirements.txt
open-comps
fastapi